requirements.txt
bottle==0.12.25
simplejson==3.19.2
ruamel.yaml==0.18.6
